This actually contains the work for prepare stop reels command moved the stop all animation tags to the lights test page

+++ Gaming/Aristocrat.Monaco.Gaming/Commands/StopAllAnimationTagsCommandHandler.cs
@@ -32,7 +32,7 @@
             }
 
             var success = _animationCapabilities.StopAllAnimationTags(
-                (int)_animationCapabilities.AnimationFiles
+                _animationCapabilities.AnimationFiles
                     .First( x => x .FriendlyName == command.AnimationName )
                     .AnimationId
                 ).Result;
